ref: move send email to utils
