--- server/app/email/index.ts
@@ -1,1 +1,0 @@
-export { sendEmail } from './send-email'
