--- server/app/email/types.ts
@@ -1,10 +1,0 @@
-export interface EmailTemplate {
-  text: string
-  html: string
-}
-
-export interface EmailSenderPayload {
-  to: string
-  subject: string
-  template: EmailTemplate
-}
